Major refactor to allow for link builders
Link builders register a handler with the main widget. When called, the
handler is given person data and returns a link object or a list of link
objects which are processed by the main widget object to create links in
the necessary places. This makes it easier to add links for other sites
in the future, cleans up the code, and will make it nicer for me to add
search widgets to places other than the person page.